remove big deps
