+++ bower.json
@@ -2,15 +2,10 @@
   "name": "sqwidget",
   "version": "0.0.1",
   "dependencies": {
-    "underscore-amd" : "",
-    "backbone-amd" : "",
     "requirejs": "",
     "requirejs-i18n": "",
-    "jquery": "",
-    "require-css": "",
-    "requirejs-text": "",
-    "requirejs-ractive": "",
-    "ractive": "",
-    "normalize-css": ""
+    "bonzo": "~1.3.6",
+    "heir": "~2.0.0",
+    "qwery": "~3.4.1"
   }
 }
