+++ src/sqwidget.coffee
@@ -1,12 +1,14 @@
 requirejs.config
   paths:
-    jquery: '../../lib/jquery/jquery'
-    underscore: '../../lib/underscore-amd/underscore'
-    backbone: '../../lib/backbone-amd/backbone'
+    ondomready: '../../lib/ondomready'
+    qwery: '../../lib/qwery/qwery'
+    heir: '../../lib/heir/heir'
+    EventEmitter: '../../node_modules/wolfy87-eventemitter/EventEmitter'
+    bonzo: '../../lib/bonzo/bonzo'
 
 # The module that is loaded first
 requirejs [
-  'jquery'
+  'ondomready'
   'component/core'
   'component/test'
 ], ($, Core) ->
@@ -14,7 +16,7 @@ requirejs [
   # the only global object that we will use.
   sqwidget = window.sqwidget = new Core()
   # Iterate all elements and register
-  $(document).ready ->
+  ondomready ->
     $('div[data-sqwidget]').each (index) ->
       sqwidget.register(@)
