+++ widgets/test/src/index.js
@@ -1,0 +1,16 @@
+define([], function() {
+  var module = {
+    views: {}
+  };
+
+  module.Controller = (function() {
+
+    function Controller(opts) {
+      opts.el.append("<div>TEST</div>");
+    }
+
+    return Controller;
+
+  })();
+  return module;
+});
